requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.6
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.1
faiss-cpu==1.13.2
filelock==3.25.2
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.18.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
ollama==0.6.1
packaging==26.0
pandas==2.3.3
parso==0.8.6
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
protobuf==6.33.6
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.3.32
requests==2.33.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
